Slim down components manifest ref handling
Remove the unused JSON-Ref resolver/dereferencer (no in-repo consumer), simplify the
snapshot loader to a direct typed read, and rename json-references.ts to
components-ref-manifest.ts. Dedup the per-operation story-index walk in manifests.ts by
threading manifest entries through, and inline the trivial fs wrappers.
